+++ tests/conftest.py
@@ -1,110 +1,77 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
+"""Place for fixtures and configuration that will be used in most of the tests.
+A nice option is to put your ``autouse`` fixtures here.
+Functions that can be imported and re-used are more suitable for the ``helpers`` file.
 """
-    Dummy conftest.py for custom_extension.
-
-    If you don't know what this is for, just leave it empty.
-    Read more about conftest.py under:
-    https://pytest.org/latest/plugins.html
-"""
-import builtins
+import logging
 import os
-import shlex
-import stat
-from contextlib import contextmanager
-from shutil import rmtree
 
 import pytest
+from pyscaffold.log import ReportFormatter
 
-
-def set_writable(func, path, exc_info):
-    if not os.access(path, os.W_OK):
-        os.chmod(path, stat.S_IWUSR)
-        func(path)
-    else:
-        raise RuntimeError
+from .helpers import rmpath, uniqstr
 
 
 @pytest.fixture
-def tmpfolder(tmpdir):
+def tmpfolder(tmp_path):
     old_path = os.getcwd()
-    newpath = str(tmpdir)
-    os.chdir(newpath)
+    new_path = tmp_path / uniqstr()
+    new_path.mkdir(parents=True, exist_ok=True)
+    os.chdir(str(new_path))
     try:
-        yield tmpdir
+        yield new_path
     finally:
         os.chdir(old_path)
-        rmtree(newpath, onerror=set_writable)
+        rmpath(new_path)
 
 
 @pytest.fixture
-def venv(virtualenv):
-    """Create a virtualenv for each test"""
-    return virtualenv
-
-
-@pytest.fixture
-def venv_run(venv):
-    """Run a command inside the venv"""
-
-    def _run(*args, **kwargs):
-        # pytest-virtualenv doesn't play nicely with external os.chdir
-        # so let's be explicit about it...
-        kwargs["cd"] = os.getcwd()
-        kwargs["capture"] = True
-        if len(args) == 1 and isinstance(args[0], str):
-            args = shlex.split(args[0])
-        return venv.run(args, **kwargs).strip()
-
-    return _run
-
-
-@contextmanager
-def disable_import(prefix):
-    """Avoid packages being imported
-
-    Args:
-        prefix: string at the beginning of the package name
-    """
-    realimport = builtins.__import__
-
-    def my_import(name, *args, **kwargs):
-        if name.startswith(prefix):
-            raise ImportError
-        return realimport(name, *args, **kwargs)
-
-    try:
-        builtins.__import__ = my_import
-        yield
-    finally:
-        builtins.__import__ = realimport
-
-
-@pytest.fixture
-def nocookiecutter_mock():
-    with disable_import("cookiecutter"):
-        yield
-
-
-@pytest.fixture
-def cookiecutter_config(tmpfolder):
+def cookiecutter_config(tmpfolder, monkeypatch):
     # Define custom "cache" directories for cookiecutter inside a temporary
     # directory per test.
     # This way, if the tests are running in parallel, each test has its own
     # "cache" and stores/removes cookiecutter templates in an isolated way
     # avoiding inconsistencies/race conditions.
     config = (
-        'cookiecutters_dir: "{dir}/custom-cookiecutters"\n'
-        'replay_dir: "{dir}/cookiecutters-replay"'
-    ).format(dir=str(tmpfolder))
+        f'cookiecutters_dir: "{tmpfolder}/custom-cookiecutters"\n'
+        f'replay_dir: "{tmpfolder}/cookiecutters-replay"'
+    )
 
-    tmpfolder.mkdir("custom-cookiecutters")
-    tmpfolder.mkdir("cookiecutters-replay")
+    (tmpfolder / "custom-cookiecutters").mkdir(exist_ok=True, parents=True)
+    (tmpfolder / "cookiecutters-replay").mkdir(exist_ok=True, parents=True)
 
-    config_file = tmpfolder.join("cookiecutter.yaml")
-    config_file.write(config)
-    os.environ["COOKIECUTTER_CONFIG"] = str(config_file)
+    config_file = tmpfolder / "cookiecutter.yaml"
+    config_file.write_text(config)
+    monkeypatch.setenv("COOKIECUTTER_CONFIG", str(config_file))
 
-    yield
+    yield config_file
 
-    del os.environ["COOKIECUTTER_CONFIG"]
+
+@pytest.fixture
+def isolated_log(request, monkeypatch, caplog):
+    """See isolated_logger in pyscaffold/tests/conftest.py to see why this fixture
+    is important to guarantee tests checking logs work as expected.
+    This just work for multiprocess environments, not multithread.
+    """
+    if "original_logger" in request.keywords:
+        yield caplog
+        return
+
+    # Get a fresh new logger, not used anywhere
+    raw_logger = logging.getLogger(uniqstr())
+    raw_logger.setLevel(logging.NOTSET)
+    new_handler = logging.StreamHandler()
+
+    patches = {
+        "propagate": True,  # <- needed for caplog
+        "nesting": 0,
+        "wrapped": raw_logger,
+        "handler": new_handler,
+        "formatter": ReportFormatter(),
+    }
+    for key, value in patches.items():
+        monkeypatch.setattr(f"pyscaffold.log.logger.{key}", value)
+
+    try:
+        yield caplog
+    finally:
+        new_handler.close()