+++ tests/test_cookiecutter.py
@@ -1,22 +1,21 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
 import logging
 import re
-import sys
 from configparser import ConfigParser
-from os.path import exists as path_exists
+from pathlib import Path
 
 import pytest
-
-from pyscaffold.api import create_project
-from pyscaffold.cli import parse_args, process_opts, run
-from pyscaffold.templates import setup_py
+from pyscaffold import __version__ as pyscaffold_version
+from pyscaffold.api import NO_CONFIG, create_project
+from pyscaffold.cli import parse_args, run
+from pyscaffold.templates import get_template
 
 from pyscaffoldext.cookiecutter.extension import (
     Cookiecutter,
     MissingTemplate,
     NotInstalled,
 )
+
+from .helpers import disable_import
 
 pytestmark = [pytest.mark.usefixtures("cookiecutter_config")]
 
@@ -25,61 +24,65 @@ PROJ_NAME = "proj"
 COOKIECUTTER_URL = "https://github.com/pyscaffold/cookiecutter-pypackage.git"
 COOKIECUTTER_FILES = ["proj/Makefile", "proj/.github/ISSUE_TEMPLATE.md"]
 
-# TODO: Remove workaround for PyScaffold <= 4.x, see comments on class
-FLAG = (lambda ext: getattr(ext, "xflag", ext.flag))(Cookiecutter("cookiecutter"))
+FLAG = Cookiecutter().flag
 
 
 @pytest.mark.slow
 def test_create_project_with_cookiecutter(tmpfolder):
     # Given options with the cookiecutter extension,
     opts = dict(
-        project=PROJ_NAME,
+        project_path=PROJ_NAME,
         package=PROJ_NAME,
-        version="0.0.1",
+        version=pyscaffold_version,
         cookiecutter=COOKIECUTTER_URL,
-        extensions=[Cookiecutter("cookiecutter")],
+        extensions=[Cookiecutter()],
+        config_files=NO_CONFIG,
     )
+    # NO_CONFIG: avoid extra config from dev's machine interference
 
     # when the project is created,
     create_project(opts)
 
     # then cookiecutter files should exist
-    for path in COOKIECUTTER_FILES:
-        assert path_exists(path)
+    for file in COOKIECUTTER_FILES:
+        assert Path(file).exists()
 
     # and also overwritable pyscaffold files (with the exact contents)
-    assert tmpfolder.join(PROJ_NAME).join("setup.py").read() == setup_py(opts)
+    setup_py = (tmpfolder / PROJ_NAME / "setup.py").read_text()
+    assert get_template("setup_py").safe_substitute(opts) == setup_py
 
     # and the cookiecutter configuration should be stored in setup.cfg
-    existing_setup = tmpfolder.join(PROJ_NAME).join("setup.cfg").read()
+    existing_setup = (tmpfolder / PROJ_NAME / "setup.cfg").read_text()
     cfg = ConfigParser()
     cfg.read_string(existing_setup)
-    # TODO
-    # assert cfg.get("pyscaffold", "cookiecutter") == COOKIECUTTER_URL
+    assert cfg.get("pyscaffold", "cookiecutter") == COOKIECUTTER_URL
 
 
-def test_pretend_create_project_with_cookiecutter(tmpfolder, caplog):
+def test_pretend_create_project_with_cookiecutter(tmpfolder, isolated_log):
     # Given options with the cookiecutter extension,
-    caplog.set_level(logging.INFO)
-    opts = parse_args([PROJ_NAME, "--pretend", FLAG, COOKIECUTTER_URL])
-    opts = process_opts(opts)
+    isolated_log.set_level(logging.INFO)
+    opts = parse_args([PROJ_NAME, "--no-config", "--pretend", FLAG, COOKIECUTTER_URL])
+    # --no-config: avoid extra config from dev's machine interference
 
     # when the project is created,
     create_project(opts)
 
     # then files should exist
-    assert not path_exists(PROJ_NAME)
-    for path in COOKIECUTTER_FILES:
-        assert not path_exists(path)
+    assert not Path(PROJ_NAME).exists()
+    for file in COOKIECUTTER_FILES:
+        assert not Path(file).exists()
 
     # but activities should be logged
-    logs = caplog.text
+    logs = isolated_log.text
     assert re.search(r"run.+cookiecutter", logs)
 
 
 def test_create_project_with_cookiecutter_but_no_template(tmpfolder):
     # Given options with the cookiecutter extension, but no template
-    opts = dict(project=PROJ_NAME, extensions=[Cookiecutter("cookiecutter")])
+    opts = dict(
+        project_path=PROJ_NAME, extensions=[Cookiecutter()], config_files=NO_CONFIG
+    )
+    # NO_CONFIG: avoid extra config from dev's machine interference
 
     # when the project is created,
     # then an exception should be raised.
@@ -89,23 +92,26 @@ def test_create_project_with_cookiecutter_but_no_template(tmpfolder):
 
 def test_create_project_without_cookiecutter(tmpfolder):
     # Given options without the cookiecutter extension,
-    opts = dict(project=PROJ_NAME)
+    opts = dict(project_path=PROJ_NAME, config_files=NO_CONFIG)
+    # NO_CONFIG: avoid extra config from dev's machine interference
 
     # when the project is created,
     create_project(opts)
 
     # then cookiecutter files should not exist
-    for path in COOKIECUTTER_FILES:
-        assert not path_exists(path)
+    for file in COOKIECUTTER_FILES:
+        assert not Path(file).exists()
 
 
-def test_create_project_no_cookiecutter(tmpfolder, nocookiecutter_mock):
+@disable_import("cookiecutter")
+def test_create_project_no_cookiecutter(tmpfolder):
     # Given options with the cookiecutter extension,
     # but without cookiecutter being installed,
     opts = dict(
-        project=PROJ_NAME,
+        project_path=PROJ_NAME,
         cookiecutter=COOKIECUTTER_URL,
-        extensions=[Cookiecutter("cookiecutter")],
+        extensions=[Cookiecutter()],
+        config_files=NO_CONFIG,
     )
 
     # when the project is created,
@@ -116,15 +122,15 @@ def test_create_project_no_cookiecutter(tmpfolder, nocookiecutter_mock):
 
 def test_create_project_cookiecutter_and_update(tmpfolder, capsys):
     # Given a project exists
-    create_project(project=PROJ_NAME)
+    create_project(project_path=PROJ_NAME, config_files=NO_CONFIG)
 
     # when the project is updated
     # with the cookiecutter extension,
     opts = dict(
-        project=PROJ_NAME,
+        project_path=PROJ_NAME,
         update=True,
         cookiecutter=COOKIECUTTER_URL,
-        extensions=[Cookiecutter("cookiecutter")],
+        extensions=[Cookiecutter()],
     )
     create_project(opts)
 
@@ -139,24 +145,26 @@ def test_create_project_cookiecutter_and_update(tmpfolder, capsys):
 @pytest.mark.slow
 def test_cli_with_cookiecutter(tmpfolder):
     # Given the command line with the cookiecutter option,
-    sys.argv = ["pyscaffold", PROJ_NAME, FLAG, COOKIECUTTER_URL]
+    args = ["--no-config", PROJ_NAME, FLAG, COOKIECUTTER_URL]
+    # --no-config: avoid extra config from dev's machine interference
 
     # when pyscaffold runs,
-    run()
+    run(args)
 
     # then cookiecutter files should exist
-    for path in COOKIECUTTER_FILES:
-        assert path_exists(path)
+    for file in COOKIECUTTER_FILES:
+        assert Path(file).exists()
 
 
 def test_cli_with_cookiecutter_but_no_template(tmpfolder, capsys):
     # Given the command line with the cookiecutter option, but no template
-    sys.argv = ["pyscaffold", PROJ_NAME, FLAG]
+    args = ["--no-config", PROJ_NAME, FLAG]
+    # --no-config: avoid extra config from dev's machine interference
 
     # when pyscaffold runs,
     # then an exception should be raised.
     with pytest.raises(SystemExit):
-        run()
+        run(args)
 
     # make sure the exception is related to the missing argument
     out, err = capsys.readouterr()
@@ -165,24 +173,26 @@ def test_cli_with_cookiecutter_but_no_template(tmpfolder, capsys):
 
 def test_cli_without_cookiecutter(tmpfolder):
     # Given the command line without the cookiecutter option,
-    sys.argv = ["pyscaffold", PROJ_NAME]
+    args = ["--no-config", PROJ_NAME]
+    # --no-config: avoid extra config from dev's machine interference
 
     # when pyscaffold runs,
-    run()
+    run(args)
 
     # then cookiecutter files should not exist
-    for path in COOKIECUTTER_FILES:
-        assert not path_exists(path)
+    for file in COOKIECUTTER_FILES:
+        assert not Path(file).exists()
 
 
 def test_cli_with_cookiecutter_and_update(tmpfolder, capsys):
     # Given a project exists
-    create_project(project=PROJ_NAME)
+    create_project(project_path=PROJ_NAME, config_files=NO_CONFIG)
 
     # when the project is updated
     # with the cookiecutter extension,
-    sys.argv = ["pyscaffold", PROJ_NAME, "--update", FLAG, COOKIECUTTER_URL]
-    run()
+    args = ["--no-config", PROJ_NAME, "--update", FLAG, COOKIECUTTER_URL]
+    # --no-config: avoid extra config from dev's machine interference
+    run(args)
 
     # then a warning should be displayed
     out, err = capsys.readouterr()