+++ tests/travis_install.sh
@@ -25,9 +25,8 @@ if [[ "$DISTRIB" == "conda" ]]; then
 
         # Use the miniconda installer for faster download / install of conda
         # itself
-        wget http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh \
-            -O miniconda.sh
-        chmod +x miniconda.sh && ./miniconda.sh -b -p $HOME/miniconda
+        wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
+        chmod +x miniconda.sh && ./miniconda.sh -b -p "$HOME/miniconda"
     fi
     export PATH=$HOME/miniconda/bin:$PATH
     # Make sure to use the most updated version
@@ -42,7 +41,7 @@ fi
 
 # for all
 pip install -U pip setuptools
-pip install tox
+pip install tox mypy
 
 if [[ "$COVERAGE" == "true" ]]; then
     pip install -U pytest-cov pytest-virtualenv coverage coveralls flake8 pre-commit