+++ CHANGELOG.rst
@@ -2,7 +2,10 @@
 Changelog
 =========
 
-Version 0.1
-===========
+Version 0.1 (development)
+========================
 
-- Initial release
+- Initial release extracted from PyScaffold
+- Updated to PyScaffold v4
+- Project reformatted using Black
+- Updated parameters passed down to Cookiecutter